refactor chunk creation to layer, disambiguating
